+++ app/src/main/java/com/android/shaftschematic/ui/screen/ShaftRoute.kt
@@ -194,6 +194,7 @@ fun ShaftRoute(
         onSelectComponentById = vm::selectComponentById,
 
         onAddBody   = { s, l, d, kwW, kwD, kwL, kwO, kwEnd, kwSp -> vm.addBodyAt(s, l, d, kwW, kwD, kwL, kwO, kwEnd, kwSp) },
+        onSetAutoBodyDia = vm::setAutoBodyDiaMm,
         onAddTaper  = { s, l, sd, ed, rate, kwW, kwD, kwL, kwO, kwS -> vm.addTaperAt(s, l, sd, ed, rate, kwW, kwD, kwL, kwO, kwS) },
         onAddThread = { s, l, maj, p, ex, aft -> vm.addThreadAt(s, l, maj, p, ex, aft) },
         onAddLiner  = { s, l, od, ref -> vm.addLinerAt(s, l, od, ref) },